VER3-218 move valid flag to study-form. it doesn't belong in study service. remove targetEvent and solveFor properties ouff study-form and into studyDesign object where they belong
