+++ src/app/shared/study-design.ts
@@ -1,18 +1,13 @@
 export class StudyDesign {
-  private _guided: boolean;
   private _name: string;
+  private _targetEvent: string;
+  private _solveFor: string;
 
   constructor(name?: string,
-              guided?: boolean
+              guided?: boolean,
+              targetEvent?: string,
+              solveFor?: string
 ) {}
-
-  get guided(): boolean {
-    return this._guided;
-  }
-
-  set guided(value: boolean) {
-    this._guided = value;
-  }
 
   get name(): string {
     return this._name;
@@ -20,5 +15,21 @@ export class StudyDesign {
 
   set name(value: string) {
     this._name = value;
+  }
+
+  get targetEvent(): string {
+    return this._targetEvent;
+  }
+
+  set targetEvent(value: string) {
+    this._targetEvent = value;
+  }
+
+  get solveFor(): string {
+    return this._solveFor;
+  }
+
+  set solveFor(value: string) {
+    this._solveFor = value;
   }
 }
